frontend/requirements.txt
# Streamlit Frontend Dependencies
streamlit>=1.40.0
requests>=2.32.0
plotly>=5.24.0
pandas>=2.2.0
httpx>=0.28.0
python-dotenv>=1.0.0
